Added Loss Tests
Added loss tests for the multiple types of Classification and Regression losses.

Changed the zero difference tensor for the Smooth L1 loss. Added testing for the multibox loss and its possible scenarios.

Integrated the AMP precision config into the functions to allow for easier mixed precision in the execution on the NVIDIA chips on AWS EC2